requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

imageio
matplotlib

# data science
numpy == 1.19.4
pandas == 1.1.5
scikit-learn == 0.24.0
tensorflow-addons

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six>=1.14
joblib
memoized-property
termcolor
pathlib